requirements.txt
requests
python-dotenv
pymongo
openai
fastapi
pydantic
apscheduler
uvicorn